+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [#3289](https://github.com/influxdb/influxdb/issues/3289): InfluxDB crashes on floats without decimal
 - [#3298](https://github.com/influxdb/influxdb/pull/3298): Corrected WAL & flush parameters in default config. Thanks @jhorwit2
 - [#3152](https://github.com/influxdb/influxdb/issues/3159): High CPU Usage with unsorted writes
+- [#3307](https://github.com/influxdb/influxdb/pull/3307): Fix regression parsing boolean values True/False
 
 ## v0.9.1 [2015-07-02]